requirements.txt
torch==2.2.2
torchvision==0.17.2

facenet-pytorch==2.5.2

opencv-python==4.8.1.78
numpy==1.26.4
pillow==10.3.0
tqdm==4.67.1
requests>=2.31.0

matplotlib>=3.2.2
pandas>=1.1.4
seaborn>=0.11.0
PyYAML>=5.3.1
